+++ classlib/src/main/java/org/teavm/classlib/java/util/TReversedLinkedHashMap.java
@@ -17,7 +17,6 @@ package org.teavm.classlib.java.util;
 
 import java.util.function.BiConsumer;
 import java.util.function.BiFunction;
-import java.util.function.Consumer;
 import java.util.function.Function;
 
 class TReversedLinkedHashMap<K, V> extends TAbstractMap<K, V> implements TSequencedMap<K, V> {
@@ -209,367 +208,16 @@ class TReversedLinkedHashMap<K, V> extends TAbstractMap<K, V> implements TSequen
 
     @Override
     public TSequencedSet<K> sequencedKeySet() {
-        return new LinkedHashMapKeySet<>(base, true);
+        return new TLinkedHashMapKeySet<>(base, true);
     }
 
     @Override
     public TSequencedCollection<V> sequencedValues() {
-        return new LinkedHashMapValues<>(base, true);
+        return new TLinkedHashMapValues<>(base, true);
     }
 
     @Override
     public TSequencedSet<Entry<K, V>> sequencedEntrySet() {
-        return new LinkedHashMapEntrySet<>(base, true);
-    }
-
-    private static <T> T checkNotNull(T node) {
-        if (node == null) {
-            throw new TNoSuchElementException();
-        }
-        return node;
-    }
-
-    static class LinkedHashMapEntrySet<K, V> extends TAbstractSet<TMap.Entry<K, V>>
-            implements TSequencedSet<TMap.Entry<K, V>> {
-        private final TLinkedHashMap<K, V> base;
-        private final boolean reversed;
-
-        LinkedHashMapEntrySet(TLinkedHashMap<K, V> base, boolean reversed) {
-            this.base = base;
-            this.reversed = reversed;
-        }
-
-        @Override
-        public final int size() {
-            return base.elementCount;
-        }
-
-        @Override
-        public final void clear() {
-            base.clear();
-        }
-
-        @Override
-        public final TIterator<TMap.Entry<K, V>> iterator() {
-            return new EntryIterator<>(base, reversed);
-        }
-
-        @Override
-        public final boolean contains(Object o) {
-            if (o instanceof TMap.Entry) {
-                TMap.Entry<?, ?> oEntry = (TMap.Entry<?, ?>) o;
-                TMap.Entry<K, V> entry = base.entryByKey(oEntry.getKey());
-                return entry != null && TObjects.equals(entry.getValue(), oEntry.getValue());
-            }
-            return false;
-        }
-
-        @Override
-        public boolean remove(Object object) {
-            if (object instanceof TMap.Entry) {
-                TMap.Entry<?, ?> oEntry = (TMap.Entry<?, ?>) object;
-                TMap.Entry<K, V> entry = base.entryByKey(oEntry.getKey());
-                if (entry != null && TObjects.equals(entry.getValue(), oEntry.getValue())) {
-                    base.remove(entry.getKey());
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        @Override
-        public final void forEach(Consumer<? super TMap.Entry<K, V>> action) {
-            if (base.elementCount > 0) {
-                int prevModCount = base.modCount;
-                TLinkedHashMap.LinkedHashMapEntry<K, V> entry = reversed ? base.tail : base.head;
-                do {
-                    action.accept(entry);
-                    entry = reversed ? entry.chainBackward : entry.chainForward;
-                    if (base.modCount != prevModCount) {
-                        throw new TConcurrentModificationException();
-                    }
-                } while (entry != null);
-            }
-        }
-
-        @Override
-        public final void addFirst(TMap.Entry<K, V> e) {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public final void addLast(TMap.Entry<K, V> e) {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public final TMap.Entry<K, V> getFirst() {
-            return checkNotNull(reversed ? base.tail : base.head);
-        }
-
-        @Override
-        public final TMap.Entry<K, V> getLast() {
-            return checkNotNull(reversed ? base.head : base.tail);
-        }
-
-        @Override
-        public final TMap.Entry<K, V> removeFirst() {
-            var e = checkNotNull(reversed ? base.tail : base.head);
-            base.remove(e.key);
-            return e;
-        }
-
-        @Override
-        public final TMap.Entry<K, V> removeLast() {
-            var e = checkNotNull(reversed ? base.head : base.tail);
-            base.remove(e.key);
-            return e;
-        }
-
-        @Override
-        public TSequencedSet<TMap.Entry<K, V>> reversed() {
-            return new LinkedHashMapEntrySet<>(base, !reversed);
-        }
-    }
-
-    static class LinkedHashMapKeySet<K> extends TAbstractSet<K> implements TSequencedSet<K> {
-        private final TLinkedHashMap<K, ?> base;
-        private final boolean reversed;
-
-        LinkedHashMapKeySet(TLinkedHashMap<K, ?> base, boolean reversed) {
-            this.base = base;
-            this.reversed = reversed;
-        }
-
-        @Override
-        public final int size() {
-            return base.elementCount;
-        }
-
-        @Override
-        public final void clear() {
-            base.clear();
-        }
-
-        @Override
-        public final TIterator<K> iterator() {
-            return new KeyIterator<>(base, reversed);
-        }
-
-        @Override
-        public final boolean contains(Object o) {
-            return base.containsKey(o);
-        }
-
-        @Override
-        public final boolean remove(Object key) {
-            int befCount = base.elementCount;
-            base.remove(key);
-            return base.elementCount != befCount;
-        }
-
-        @Override
-        public final void forEach(Consumer<? super K> action) {
-            if (base.elementCount > 0) {
-                int prevModCount = base.modCount;
-                TLinkedHashMap.LinkedHashMapEntry<K, ?> entry = reversed ? base.tail : base.head;
-                do {
-                    action.accept(entry.key);
-                    entry = reversed ? entry.chainBackward : entry.chainForward;
-                    if (base.modCount != prevModCount) {
-                        throw new TConcurrentModificationException();
-                    }
-                } while (entry != null);
-            }
-        }
-
-        @Override
-        public final K getFirst() {
-            return checkNotNull(reversed ? base.tail : base.head).key;
-        }
-
-        @Override
-        public final K getLast() {
-            return checkNotNull(reversed ? base.head : base.tail).key;
-        }
-
-        @Override
-        public final K removeFirst() {
-            var e = checkNotNull(reversed ? base.tail : base.head);
-            base.remove(e.key);
-            return e.key;
-        }
-
-        @Override
-        public final K removeLast() {
-            var e = checkNotNull(reversed ? base.head : base.tail);
-            base.remove(e.key);
-            return e.key;
-        }
-
-        @Override
-        public TSequencedSet<K> reversed() {
-            if (reversed) {
-                return base.sequencedKeySet();
-            } else {
-                return new LinkedHashMapKeySet<>(base, true);
-            }
-        }
-    }
-
-    static class LinkedHashMapValues<V> extends TAbstractCollection<V> implements TSequencedCollection<V> {
-        private final TLinkedHashMap<?, V> base;
-        private final boolean reversed;
-
-        LinkedHashMapValues(TLinkedHashMap<?, V> base, boolean reversed) {
-            this.base = base;
-            this.reversed = reversed;
-        }
-
-        @Override
-        public final int size() {
-            return base.size();
-        }
-
-        @Override
-        public final void clear() {
-            base.clear();
-        }
-
-        @Override
-        public final TIterator<V> iterator() {
-            return new ValueIterator<>(base, reversed);
-        }
-
-        @Override
-        public final boolean contains(Object o) {
-            return base.containsValue(o);
-        }
-
-        @Override
-        public final void forEach(Consumer<? super V> action) {
-            if (base.elementCount > 0) {
-                int prevModCount = base.modCount;
-                TLinkedHashMap.LinkedHashMapEntry<?, V> entry = reversed ? base.tail : base.head;
-                do {
-                    action.accept(entry.value);
-                    entry = reversed ? entry.chainBackward : entry.chainForward;
-                    if (base.modCount != prevModCount) {
-                        throw new TConcurrentModificationException();
-                    }
-                } while (entry != null);
-            }
-        }
-
-        @Override
-        public final V getFirst() {
-            return checkNotNull(reversed ? base.tail : base.head).value;
-        }
-
-        @Override
-        public final V getLast() {
-            return checkNotNull(reversed ? base.head : base.tail).value;
-        }
-
-        @Override
-        public final V removeFirst() {
-            THashMap.HashEntry<?, V> e = checkNotNull(reversed ? base.tail : base.head);
-            return base.remove(e.key);
-        }
-
-        @Override
-        public final V removeLast() {
-            THashMap.HashEntry<?, V> e = checkNotNull(reversed ? base.head : base.tail);
-            return base.remove(e.key);
-        }
-
-        @Override
-        public TSequencedCollection<V> reversed() {
-            if (reversed) {
-                return base.sequencedValues();
-            } else {
-                return new LinkedHashMapValues<>(base, true);
-            }
-        }
-    }
-
-    private static class EntryIterator<K, V> extends AbstractMapIterator<K, V> implements TIterator<Entry<K, V>> {
-        EntryIterator(TLinkedHashMap<K, V> map, boolean reversed) {
-            super(map, reversed);
-        }
-
-        @Override
-        public Entry<K, V> next() {
-            makeNext();
-            return currentEntry;
-        }
-    }
-
-    private static class KeyIterator<K, V> extends AbstractMapIterator<K, V> implements TIterator<K> {
-        KeyIterator(TLinkedHashMap<K, V> map, boolean reversed) {
-            super(map, reversed);
-        }
-
-        @Override
-        public K next() {
-            makeNext();
-            return currentEntry.key;
-        }
-    }
-
-    private static class ValueIterator<K, V> extends AbstractMapIterator<K, V> implements TIterator<V> {
-        ValueIterator(TLinkedHashMap<K, V> map, boolean reversed) {
-            super(map, reversed);
-        }
-
-        @Override
-        public V next() {
-            makeNext();
-            return currentEntry.value;
-        }
-    }
-
-    static class AbstractMapIterator<K, V> {
-        private final TLinkedHashMap<K, V> base;
-        private final boolean reversed;
-        private int expectedModCount;
-        private TLinkedHashMap.LinkedHashMapEntry<K, V> futureEntry;
-        TLinkedHashMap.LinkedHashMapEntry<K, V> currentEntry;
-
-        AbstractMapIterator(TLinkedHashMap<K, V> base, boolean reversed) {
-            this.base = base;
-            this.reversed = reversed;
-            expectedModCount = base.modCount;
-            futureEntry = reversed ? base.tail : base.head;
-        }
-
-        public boolean hasNext() {
-            return futureEntry != null;
-        }
-
-        final void checkConcurrentMod() throws TConcurrentModificationException {
-            if (expectedModCount != base.modCount) {
-                throw new TConcurrentModificationException();
-            }
-        }
-
-        final void makeNext() {
-            checkConcurrentMod();
-            if (!hasNext()) {
-                throw new TNoSuchElementException();
-            }
-            currentEntry = futureEntry;
-            futureEntry = reversed ? futureEntry.chainBackward : futureEntry.chainForward;
-        }
-
-        public void remove() {
-            if (currentEntry == null) {
-                throw new IllegalStateException();
-            }
-            checkConcurrentMod();
-            base.remove(currentEntry.key);
-            currentEntry = null;
-            expectedModCount++;
-        }
+        return new TLinkedHashMapEntrySet<>(base, true);
     }
 }
