+++ classlib/src/main/java/org/teavm/classlib/java/util/function/TPredicate.java
@@ -37,6 +37,7 @@ public interface TPredicate<T> {
         return t -> TObjects.equals(t, targetRef);
     }
 
+    @SuppressWarnings("unchecked")
     static <T> TPredicate<T> not(TPredicate<? super T> target) {
         return (TPredicate<T>) target.negate();
     }
